Updated openstack/openstack Project: openstack/sahara  b0f06e740edcd0127f73a024f62f892673f66d98
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Icc0f1c76e7d86336f4254d22e36ee5006d8e7b27