Updated openstack/openstack Project: openstack/cinder  53c93a6ec35270e2b892b3bbdf7b915a95c31d64
VMAX driver - Replace "SLO" tag with "ServiceLevel" tag

The VMAX3 XML file has a tag currently named "SLO"
which we will replace with "ServiceLevel", to be
consistent with terminology used by VMAX3 community,
as well as the CamelCase naming convention. This patch
modifies the code to reflect the recommended change in
the XML file.

Change-Id: I658fc1c618607995d724d3df89f2055da767670e
Closes-Bug: #1618908