Updated openstack/openstack Project: openstack/oslo.vmware  15f00899462846f34b50681b46774f8ec13027f9
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I8f33aab4e10fb8905a93ec772f07117a1c8955bf