Updated openstack/openstack Project: openstack/cinder  c3831f176c23d936054f3f5dfde587e709865dec
VMAX driver - Heat detach issue.

There is a threading issue in Heat on a delete stack. All
detaches happen simultaneously. The fix locks the portion
of code to determine accurately, the number of volumes left
in the storage group, at any given time, to prevent
attempting to remove a volume that is the last in the
storage group.

Change-Id: I84164a9abac35d962408febbe8d3af759beb6a94
Closes-Bug: #1630535