Updated openstack/openstack Project: openstack/python-openstackclient  5ba5e1e321712578029c3de499d2992a955e2d11
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Ib265f919e176f048f6e56bde47db3d99b3dd993e