Updated openstack/openstack Project: openstack/python-openstackclient  7e7f75575424fe6d88cd4005acef2ae9a8b5ecbf
Mask passwords in debug logs for auth_config_hook

The auth config hook can have credentials in it so
we have to mask the config before logging it. To
avoid doing the work of masking the password if we
aren't going to log it, there is a conditional put
around the actual debug statement.

Change-Id: I8e626672ec94fc837610216bccb4354dbdedca17
Closes-Bug: #1630822