Updated openstack/openstack Project: openstack/neutron  7311e56c490e4b8f1fce07996c898b622e42206b
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Ic0d37f26c59aeed3a55a5b229404c1d18bf248f0