Updated openstack/openstack Project: openstack/barbican  df4664fec333ffb8ce42357e570ae70f9e0a7b60
Fix routing for adding a secret to a container

This CR fixes the bug where sending a POST to add a secret to a
container would fail with 404 - Not Found when the path included a
traling forward slash (/).

The underlying issue was that the remainder tuple returned by the pecan
framework is of different lengths depending on whether the url includes
the trailing slash or not.

Change-Id: I89282ad78ceedf4db795564f59306b9e88cc0612
Closes-Bug: 1627176