Updated openstack/openstack Project: openstack/heat  ed649dd32b622a11e56cdd1c1af6153b08da7f5b
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I4b94ae2d64e8660e1be4f25cf009fb98f550fbb3