Updated openstack/openstack Project: openstack-infra/project-config  79913bbda894b574db9a28b92acdb5b12135c0cd
Add check-tripleo-undercloud-jobs template

Adds a new template, check-tripleo-undercloud-jobs, and moves the
undercloud and undercloud-upgrades jobs to the template. Updates the
instack-undercloud and python-tripleoclient projects to use the new
template.

This means that other projects using only the check-tripleo-jobs
template will no longer run the undercloud job, but that functionality
is wholly contained with the multinode job, so no loss of coverage is
introduced.

Change-Id: Ia708f65ad7096593512424f2a5ca4d5906c32e79