Updated openstack/openstack Project: openstack/diskimage-builder  5897e4975d1767bc20917436b4683985e8b5c189
Disabling all previous repos registered in the system

Depending on the pool id used, so many repos are brought,
including not valid ones that cause image to crash, or repos
that include conflicting packages.
Before enabling repos, disable all previous ones, so we
can be sure that we only bring the repos specified in the
parameters.

Change-Id: Ifd4d8d1d4fa954cd2593669e516e3201f2d6f6c1