Updated openstack/openstack Project: openstack/glance  bbe47c32b949f98a982200cb56ed6d10b57bb4eb
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I3c96dc9ecd25f499d826542f47f8d65b211839ca