Updated openstack/openstack Project: openstack/neutron  365479949651f996c3d20d80ed6a3f884c5cec99
Fix misleading error trace in trunk fullstack life cycle test

There is a race condition triggered by the test where a VM and subports
associated to trunk linked to the VM are deleted in fast sequence. What
happens is that subport removal can no longer store metadata associated
to the trunk as the trunk is being removed.

Even though the race does not lead to errors, the trace may mislead the
operator in thinking there is something worth investigating when in reality
everything is fine.

To solve this issue, we should avoid raising RuntimeError for when the
parent port is not found, because that prevents us from handling this
race correctly.

Related-blueprint: vlan-aware-vms

Change-Id: Ica757813fd0348235cc7b823eebf6d3588c79905