Updated openstack/openstack Project: openstack/nova  152d5e45ef698f0a9f8b05a834086c3a973ee527
Delete traces of in-progress snapshot on VM being deleted

When user tries to create snapshot of instance and at the same time
if another request tries to delete the instance, at that time image
goes in saving status for forever because of race  between instance
delete and snapshot requests.

Caught exceptions(InstanceNotFound and UnexpectedDeletingTaskStateError)
in except block and deleting the image which got stuck in saving state.

Closes-Bug: #1555065
Change-Id: If0b918dc951030e6b6ffba147443225e0e4a370a