Updated openstack/openstack Project: openstack/neutron  552adc9e1c77cb4d805e27c8a34634c2bc44b257
Deterministic ordering of fixed_ips

This adds an order_by clause to the fixed_ips relationship
on the port object to ensure that the fixed_ip ordering is
consistent between a create, an update, and a get request
for a port. Without it we were at the mercy of the sql backend
to determine how it felt like ordering them on the join condition.

Closes-Bug: #1630410
Change-Id: I523e0ab6e376f5ff6205b1cc1748aa6d546919cb