Updated openstack/openstack Project: openstack/glance_store  40786d508e7605c0bb950e07d8698e79546a9f11
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I6e979324198853b728a18b9271af1357d4986b1f