Updated openstack/openstack Project: openstack/nova  1feda57285a5c344f098ee9c77c0278aecc9e151
[placement] Allow both /placement and /placement/ to work

When mounted under a prefix (such as /placement) the service was
only returning the home document at /placement/ not /placement. In
the context of having a prefix, we'd generally like the latter to
work and for '/' to work when there is no prefix. This allows both.

Note that this doesn't make /placement/resource_providers/ work, and
we don't want that.

Change-Id: I0ac92bf9982227d5f4915175182e5230aeb039b4
Closes-Bug: #1626490