Updated openstack/openstack Project: openstack/nova  38c4516f8a77a689f1835f041cd69b3f08acabda
Change log level to debug for migrations pairing

For resize/cold-migration it's possible that instance
already changed host to destination, but no confirm/revert
has happened yet. In that case resource tracked starts spamming
errors, because it's impossible to match migration and instance.
It's safe to lower log level to debug in that case.

Change-Id: I70cb7426e0e2849ee7d01205ce7b2d883a126d66
Closes-Bug: #1555320