Updated openstack/openstack Project: openstack/python-glanceclient  5eba376eed36beb42be43abd9eb38eda03bffa4f
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Ifad58c69d888e58cf1cc998bb1ddb409b4489490