Updated openstack/openstack Project: openstack/nova  948f57d832c7a7c07ab7fcd736dac941915b8346
Annotate online db migrations with cycle added

When looking at removing online migrations it's helpful to know which
cycle they were added in. This helps ensure that migrations aren't
removed in the same cycle they're added, and makes it easy to spot a
migration that was added a long time ago and may be ripe for removal.

Change-Id: Ibc7b0c5ab29389553f31cde52c5f5a6fc43119f2