Updated openstack/openstack Project: openstack/diskimage-builder  3c1f82df4618d6682464008c80268d075c3ae352
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Ib3cfb6a2cc014b72c32db7434975d4bb75d480d7