Updated openstack/openstack Project: openstack/zaqar  0af1fe3e0bb6a0f9d5859877d038f67fa3fe4926
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I7764d9086884c046c8d0585a0e8ede6e860170e1