Updated openstack/openstack Project: openstack-infra/project-config  1a0663f9ae87eb7316878ae285f6dab767b556df
Remove propose-puppet-openstack-rdo-promote

propose-puppet-openstack-rdo-promote was useful to submit changes every
night to promote repositories in Puppet OpenStack CI when deploying with
RDO.

Now, we have the process automated in RDO Infra, and we can use a static
URL that is auto-promoted if RDO CI jobs pass with new packaging.

In other words:

- Before:
Puppet OpenStack was testing OpenStack for a repo older than 24h and
manually promoted repositories every day.

- Now:
Puppet OpenStack will use the same repository that is automatically
updated in RDO every 3 hours if jobs are passing.
We're improving the CI/CD toolchain.

It depends on this change in puppet-openstack-integration:
https://review.openstack.org/382940

Change-Id: I52b6098163bd95aae5d8bc69bf12499fa0ddacc9
Depends-On: Ia1ac4b55f8542d94c68dc761a9af04c447d50e3e