Updated openstack/openstack Project: openstack/openstack-manuals  31443ecd1691c6cbd70c8928d5a8c241e6d94b82
[networking] Reorganize deployment examples

Major reorganization and rewriting of deployment examples to
use "building blocks" that enable the audience to begin with
a simple architecture and add complexity as necessary. Also
reduces duplication of content.

Configuration and commands reference Mitaka to ease backporting
to Mitaka. After merge, I will submit another patch to update
the configuration and commands for Newton.

Macvtap content needs a significant rewrite to integrate
better with the deployment examples. Will address in Ocata.

Change-Id: I2ae95d726aa053a0f30507a3e68907ac56296e6b
backport: mitaka