Updated openstack/openstack Project: openstack/training-guides  73bd7ee517ecc71c3dce9dabefff3f9cf66ee30d
Reorganize directory structure

Currently for both upstream-training and training-guide directory
slides are stored in the subdirectories of landing web pages.
It is not easy to capture the directory structure and this requires
Sphinx conf file to have unnecessary excluded entries.

This commits reorganizes each directory to the following structure.
It follows the convention in the documentation project.

doc/upstream-training:

(previous)
  +-- README.rst
  +-- website RST files
  +-- conf.py (for website)
  +-- build/
  +-- source/
       +-- slide RST files
       +-- conf.py (for slides)
       +-- _assets/
       +-- _static/
       +-- locale/

(new)
  +-- README.rst
  +-- build/
  +-- source/
       +-- website/
       |    +-- website RST files
       |    +-- conf.py (for website)
       |    +-- _assets/
       +-- slides/
       |    +-- slide RST files
       |    +-- conf.py (for slides)
       |    +-- _assets/
       |    +-- _static/
       +-- locales/

training-guide directory is reorganized similarly.
training-guide _assets/README has been removed as it is used.

Also updates tox.ini to share target command definitions.
We do not need to have same things in many places.

Note that this commit does not change the build document paths.

Change-Id: I5a21f170c06d8f71f1f2d0df354e6d6280cb29d0