Updated openstack/openstack Project: openstack/python-manilaclient  023b2b3148669fa7e76c1d1d5a5e6a1f91947fde
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I492355f4dabe224152d43a5722a4459548261205