Updated openstack/openstack Project: openstack-infra/infra-manual  e5c2d669a12c8f45324b8faa3771fe86367ca38d
emphasize the prefered way for importing repository history

We have had a couple of instances this cycle where projects submitted
large numbers of patches as a way to import their project history. This
patch updates the instructions to try to emphasize why that is not a
good idea, and that history should be imported using the upstream field
instead.

Change-Id: Ia675608f66a1fe6c5c7317584e1f75f49dfb451b
Signed-off-by: Doug Hellmann <doug@doughellmann.com>