Updated openstack/openstack Project: openstack/heat  f2182e12fa0ae3510cf2a4569f3f2a9e495f2350
Add http_proxy_to_wsgi middleware to Heat CFN endpoint

This was already used in the API endpoint, but it's also needed in
the CFN endpoint. It's purpose is to process the X-Forwarded-Proto
header (or Proxy protocol if used) and set the protocol as directed
to https if done so. It's only needed if Heat is behind a TLS proxy
(such as HAProxy) and is also disabled by default.

Change-Id: Ibd81e1cf6bc1e3f63728b485e295478afa7f573c
Closes-Bug: #1590608