Updated openstack/openstack Project: openstack/keystonemiddleware  d378b11cd1cff367ed9c81896cf706106e2f34f0
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: If8131a103f01b062cca757144d06d644e5a3fec9