Updated openstack/openstack Project: openstack/ceilometer  a8066ee6f4164ef5895b0f29282a07102e3f21aa
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Icb2e410b971001e8ff225721b04d101ee8d7e3f8