Updated openstack/openstack Project: openstack/neutron-fwaas  b80de376d484e5cd1eb09df4c0720e52f3f29742
Deal with the '-m protocol' flag in iptables FwAAS v1 and v2

Iptables automatically add a '-m protocol' flag for rules containing a source
or a destination port. FwAAS do not add this flag so that, on apply, rules are
always different from iptables-save output. This induce a very long loop in
neutron-l3-agent hosting the router as the comparison of each line is just
slightly different.

This patch only add one '-m protocol' flag before port.

Closes-Bug: #1618430
Change-Id: Ia3fa3889dbf3ee10425e7e7fce8a3b8351f14e60