Updated openstack/openstack Project: openstack/manila  75c9aaf50d97b779f5cc898966010208bb473fdd
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I6d9eac9c512935a728f3d3919b1ac2f560efe403