Updated openstack/openstack Project: openstack/horizon  7b1e46856e5b064c8f9f132ef7e1a39ddf8c98ee
Using /tmp for SECRET_KEY in tests

While using openstack_dashboard/test for the path of SECRET_KEY works
perfectly for building Horizon, including when building the pacakge for
Debian, this is a no-go for Horizon plugins in the Debian context. Indeed,
ironic-ui, manila-ui and sahara-dashboard are all doing:

from openstack_dashboard.test.settings import

While this works in devstack, as it is using a clone of the Horizon git
repository which is writable, it doesn't work in the context of Debian
(or other distro) packaging. Indeed, in this case, the folder becomes
/usr/lib/python2.7/dist-packages/openstack_dashboard/test. This folder,
in all reasonableness, is owned by root, and read only by the user who
is building the package. So at import time, doing this:

SECRET_KEY = secret_key.generate_or_read_from_file(
    os.path.join(TEST_DIR, '.secret_key_store'))

just fails miserably when the horizon plugin packages are building.

Simply using /tmp instead of TEST_DIR fixes the issue. This patch makes
it in a portable way using tempfile.gettempdir() to find /tmp. Though
it isn't completely safe, as the file in /tmp is predictable (and then
there can be a symlink attack), though for unit testing we probably
don't care.

Please allow this patch to go through, so we don't have to carry it
at the distribution level.

Change-Id: Iadf0c2bffe19e2c33083a257102e846886d7680f