Updated openstack/openstack Project: openstack/nova  55db69f2c0b0a80733d22f5ca9a1b0dff5704f6d
Replace uuid4() with uuidsentinel

As of now, in most of the test cases, uuidsentinel is used for
generating a UUID except at some places where uuid4() is used.
In order to maintain consistency, we propose to use uuidsentinel
module for generating UUIDs throughout the test cases.

There are some cases where unique UUIDs are required. For such
cases, generate_uuid() from oslo_utils.uuidutils is used.

Change-Id: I00d953c29352a14508cea3a76f22cdd1a16a187a