Updated openstack/openstack Project: openstack/python-ceilometerclient  79d6e3c5ac06228fc07537356ef4e8403e3201d4
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I7c2902f64988c8b39ae8f13fec86622dcc2fb955