Updated openstack/openstack Project: openstack/oslo.messaging  c881baed29db49c5710795496cb07907e35ceaba
rabbit: Don't prefetch when batch_size is set

When the application set batch_size, so we don't need to prefetch more
messages, especially for notifications. Notifications queues can be
really big when the consumer have disapear during a long period, and
when it come back, kombu/pyamqp will fetch all messages it can. So we
override the qos prefetch value.

Change-Id: I601e10cf94310b9f96f7acb9942959aaafad7994
Closes-bug: #1551667