Updated openstack/openstack Project: openstack/barbican  b62e182d1c72633f11f22e8919d070bace800c39
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I5111eb95ceacc4900969dc60d8e67c79b805cab8