Updated openstack/openstack Project: openstack/governance  c818a13888216e1307422c6e3910fd58e6cd9dcd
Propose Thierry Carrez as TC chair for Ocata

I'd like to nominate myself to be the chair of the
Technical Committee for yet one more round. Since Doug
is still holding the PTLship for release management,
I have extra time to dedicate to the Technical Committee,
which comes handy to navigate the complex issues we need
to tackle in the near future.

Change-Id: I4bd9db2153e2ccd4873554508399fd2a02be889e