Updated openstack/openstack Project: openstack/nova  faafb6091e511a8562d6a604e25b8905532245ea
ironic: Rename private methods for instance info

This is a follow-up patch for Idf5191aa1c990552ca2340856d5d5b6ac03f7539.
This patch renames private methods to add instance information to
ironic node and remove the information for readability. A note is also
added to show why the remove method is necessary.

Change-Id: I48585bfd62ecdf0fadfbdef9e539d0b8f3d5bdc8