Updated openstack/openstack Project: openstack/keystonemiddleware  c0c8677d866309c469d25b142f4cd0b31d0489d1
Specify that unknown arguments can be passed to fetch_token

To allow flags to be added to fetch_token we need to ensure that any
implementations understand that new information can be passed via the
fetch_token function and that they should ignore this information if
they don't know how to handle it.

This just codifies this information for the equivalent keystone change.

Note that the default implementation of fetch_token in AuthProtocol has
not been updated as it should know of all flags that could be passed to
it.

Implements bp: allow-expired
Change-Id: I7312beb7cdd9527d959d6b7a94c6bfc6bf3c5952