Updated openstack/openstack Project: openstack-infra/tripleo-ci  d81b76e70878bbd00152129452b950704ca68e5c
Make the ovb-updates job work again

The updates job was never properly converted to ovb, so to my
knowledge it has never worked since the transition.  This change
should update the configuration so it will work in ovb.

Note that I also added storage-environment.yaml since previously
it appears we were deploying a ceph node but not actually using it.

Change-Id: I5b6649de5d86e5ce4004fc9667b70fb4008f7021