Updated openstack/openstack Project: openstack/tempest  6b3fbaa605e433be68eb6634619d4ddaa4df5892
Remove redundant assert judgement

The assert judgement after wait_for_qos_operations is redundant,
because the same judgement already exists in wait_for_qos_operations,
and so the latter assert judgement will never fail(an exception will
be thrown in wait_for_qos_operations if anything is wrong). so
this is to remove the redundant judgement.

Change-Id: Ia172b1747242307e0cf92941dacee5b9144852f3