Updated openstack/openstack Project: openstack/oslo.db  cfc4edcb42e4054399330e3a5933065ade52d9f5
Add a specific exception for 'unknown database' errors

Add an exception filter to distinguish 'unknown database' errors from
other DBAPIError's. This effectively means we no longer
unconditionally print a traceback to logs, when someone tries to
connect to a database, that does not exist, but just raise a specific
exception instead and let the caller handle it.

Closes-Bug: #1631033

Change-Id: Iebc1ec1e8153917fbcde116d04bc171272993bbd