Updated openstack/openstack Project: openstack/oslo.concurrency  56c27b74bb2f98cbbcf357f5c244fa9d3ed5ba50
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I6fd9f1265f05a5d4e44355def5a6519a8021e25d