Updated openstack/openstack Project: openstack/oslo.db  6153adf6440a593be8fbdb342e23b2a3df6da629
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I723a3e5f79db0e2fe0ff22795a5200391c967bff