Updated openstack/openstack Project: openstack-infra/shade  b606d25b15a48ac82d06913fcc9b2661fabbb202
Start splitting normalize functions into a mixin

As we add location information to all of the normalizers, we need to
pass in info from the cloud object. Ultimately, that means they need
access to cloud members, which means they're no longer really external
functions.

However, openstackcloud.py is super long as it is, so keep them in their
own place by making a base class to shove them in to.

Change-Id: I004e54fa26ea3745b96828100cb11dc5824416ac