Updated openstack/openstack Project: openstack/swift  0cf44fd0c695bc4900f82725766fd8e124dc85db
Update test-requirements.txt

The testing requirements should not cause extra packaging
work for deployers (since these aren't required to deploy).
This patch brings it in line with the latest global requirements,
except for python-swiftclient itself (which I kept with no version
specifier). By being in line with global requirements, this is a
more accurate list with respect to how things are actually tested
in the gate.

Change-Id: Iddc92d9bb135144d38c3e5aec2e95df5787e9b5b