Updated openstack/openstack Project: openstack/oslo.config  4a691931f378109fd9e1ec97bff0f26dde591e89
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I9db09c7585e9e1041139970ca848f77b38424a6f