Updated openstack/openstack Project: openstack-infra/shade  8d2b74fbec8d78b43d3a6769b0160a7af7b4c061
Update location info to include object owner

Many OpenStack objects (but not all, grrr) return information about the
project that owns the resource. When we add location info to an object
record, that's the information we care about. Update current_location so
that it wraps a private method that allows explicit passing in of a
project_id to override the token's project_id. If the project_id and the
token's project_id are the same, we can continue to pull out project and
domain name information for readability.

Change-Id: I2a7e6003a2b75894907f90c9ab94244a769a892f