Updated openstack/openstack Project: openstack/keystone  2102498b363847832b66fa55edf3485d21e58456
Simplify the KeystoneToken model

Now that we only really use the validate_v3_token method before
passing data to the token model, we can get rid of a lot of the
v2.0 logic. In the future, the token model shouldn't care about
any sort of version. This moves us in that direction.

Change-Id: I2687750e0c3647c3302c4479af5b2498c8d7351b