Updated openstack/openstack Project: openstack/nova  0ceccdf28801a6dad0bda7d41e35712b1ce5bbc4
doc: note the future of out of tree support

The out of tree support section of the nova dev policy
document was pretty old and basically said hooks and
extension points are buyer beware, but doesn't mention
anything about why we don't want them (interoperability)
and the efforts the last few releases to actively
deprecate and remove hooks/extension points/classloading.

This change adds some more wording to that effect to
the doc.

Change-Id: Iaa1cf3ef7ac6e8e1d75f94e17aaba1058474acca