Updated openstack/openstack Project: openstack/nova  36c459781c4f39dd4ba9db96c5ded98729a2aa62
doc: add dev policy about no new metrics monitors

At the newton midcycle we talked about the proposed
memory bandwidth monitor plugin and decided that we
really just don't want nova in the business of metrics
gathering, and are moving to deprecate that functionality.

There are several specs open related to adding more
metrics monitors and/or exposing existing ones, so this
codifies the policy on nova as metrics gatherer.

Change-Id: Ie47e91776bc36a22fd48d3673c69c842509489a1