Updated openstack/openstack Project: openstack/trove  7b1254ef439c1661a5d65668b00535a50d46f1a1
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I597e684dc1223e6292135c31e2abb56baf7e6249