Updated openstack/openstack Project: openstack/tripleo-heat-templates  75d90579973fabe8b773b4c4f4278dbf1d88a829
Special case non-matching ObjectStorage role port names

Unfortunately we use "SwiftStorage" in the ObjectStorage role
template, so we have to special-case this for backwards compatibility
or deployments enabling the ObjectStorage role will fail.

Ideally we'd align the port names in the objectstorage-role.yaml, but we
can't becauuse all the ports would be replaced in existing deployments
on update.

Change-Id: Ia07e193d2b9a4d33c6272c2b4448133584b81350
Closes-Bug: #1632663