Updated openstack/openstack Project: openstack/neutron-lbaas  b93f310861fbbb29e9fe420f0aa18cc166964dfb
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Ifcd378a2a6bb6aa77055604b461a162717bd0a91