Updated openstack/openstack Project: openstack-infra/project-config  14e7cb3eb217f9ec32c9f3a02abe03c46d631a47
Update doc to have 'debootstrap' dep and describe minimal

Update the doc, that explains how to use disk-image-builder to create an
image, to have an additional dependency on 'debootstrap'

Also remove DISTRO= line from example as it was confusing and would make
people not build the ubuntu-minimal image.

Add a little info on the -minimal build types.

Change-Id: I8a7393a9c71ef503a730e4442941996eab43da6b