Updated openstack/openstack Project: openstack/oslo.rootwrap  16c22e9b1a83c1fa27c1fac071f9c41c5822e019
Fix running unknown commands in daemon mode

Running a unknown command (which is mentioned in the filter
but not available on the filesystem) leads currently to an exception.
Make sure that the return codes for both, daemon and non-daemon mode
are equal when running the same command.
Also add functional tests for this case.

Change-Id: I20004c3c370d004b5b76f4c8f8ab167d0949fabf
Closes-Bug: #1632768