Updated openstack/openstack Project: openstack/nova  fa553be03e0338f7244ebab1923935d2a1b4f957
Renamed parameters name in config.py

In LibvirtConfigObject, the parameter's name for the
node is 'name'. Which is a problem, when we want to add
an attribute named 'name' to a DOM node. In this patchset,
the symbolic parameter name for a node's name is changed
from 'name' to 'node_name'. After this change, we can add an
attribute 'name' to our DOM node objects.

Change-Id: I7baaf7f828c3b5ffa3d0bec7e80d348f7f9d1e45