Updated openstack/openstack Project: openstack/tempest  93e6a0370085b0488b8542a66f4adcbf11f4a390
waiters.py: raise BackupException defined in tempest/lib

This patch is to fix the issue which described in the bug #1632410

A timeout is being raised on a volume backup test scenario
which could be that because the volume backup is slow or
we tripped a cinder bug. The real issue is that wait loop
for backups in tempest is referencing a non-existent exception class,
"VolumeBackupException".

Change-Id: I2d74afc29cd52da7f3642bf56ea896c92cfccdca
Closes-bug: #1632410