Updated openstack/openstack Project: openstack/cinder  858f9b5ada6eb5fc5a7412abf249cef6d603ad45
VMware: Fix deprecation warning

Currently we print a deprecation warning that the support
for vCenter 5.1 will be removed in Ocata release. But the
VMware Nova driver deprecated vCenter 5.1 only in the Ocata
release, and therefore we have to wait until the Pike release
to remove the support for vCenter 5.1. Updating the warning
message to reflect this change.

Change-Id: Ie476f4fd75d27c91bc9cb39103f9b8ec32b39056