Updated openstack/openstack Project: openstack-infra/system-config  ee085879be0625d7643c73ce2b5372f7a4e50321
Specify cacti ip addresses for snmp access

Move openstack-specific configuration into the openstack_project module.
This was previously hard-coded in the iptables module, but is now
parameterized. Before the parameter can be given a site-agnostic default
value, pass in the needed value here.

Add both the IPs for both the old and new cacti servers as part of the
service migration.

Change-Id: I173ca1efae4644c89cfab68d6beeba0a1dae9ce2
Depends-On: I9394982811f8dcf0d63eccb782de04bf4a047ec7
Needed-By: Ibae45af594fc2b18024fcc2d6ef040afd4ddd926