Updated openstack/openstack Project: openstack/openstack-manuals  2753a57882a1a1a617f73749977e057098b9030e
Add initial custom error page for new server

With the new server, we can use our own 404 server page, use it to
document to the older server (once it's setup).

this page will only be shown on the new server and should inform about
the archived content. Since the name of the archived site is not
specified yet, it uses docs.o.o.

Change-Id: I400038e52189479909720c55838dc05910e6509c