Updated openstack/openstack Project: openstack-infra/project-config  d7e41e03c2df3ef6f81ea034cef15e71889416b8
Rework translation job git setup

Since we have zuul-cloner on the proposal node now, we can use it there.

So, remove the constraints-git-prep and
zuul-git-branch-prep-upper-constraints and replace it with the normal
zuul-cloner builders that set up constraints.

The two builders have been used only for translation jobs, we can remove
them completely now.

Change-Id: I7787daf025f4f14cf665db0366fa81f54adfa884