Updated openstack/openstack Project: openstack/oslo.policy  3d34216044ff19c53a03095d8c155238b0b20acf
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: Iff0ac75910b62cb99f702c65cd8740cfbe1923a1