Updated openstack/openstack Project: openstack/oslo.messaging  211c2c1418896a4fd7669dad6a1445dc10135d77
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I7f004ff1404281a825f6317d1555bbb32822a971