Updated openstack/openstack Project: openstack-infra/project-config  ab4c3c829f92a3e6f3870afe076fd6c759458bd7
Add ansible-role-virtualenv to openstack workflow

This role will be responsible for managing virtualenv. Currently, we
are not aiming to be part of big tent. We do not have an upstream
repo, so this will be created as a new repo in gerrit.

Notifications will go to windmill channel for now.

Change-Id: Id83973b8aad871f9006a0cdf409b78d47383ff4e
Signed-off-by: Paul Belanger <pabelanger@redhat.com>