Updated openstack/openstack Project: openstack/cinder  ab8abb8aff67f6d10b1f2c167cedabc759b6c37a
NetApp NFS: Look for new file for >60s

The default lookup cache expiry time for the
Linux NFS client is 60 seconds.  This means that
looking for a new file could fail for the current
45 seconds, but if the driver looks for >60s, the
cache will be refreshed, and the file's presence
will be detected.

This means that we don't have to change the NFS
client mount options to compensate for this cache
behavior.

Change-Id: I8c63fe6b8c5de74eac3f3c259783bb946e0482a0
Closes-Bug: #1608596