Updated openstack/openstack Project: openstack/tempest  4ed7b86677b0a7d1047cb10f9cd619a80ea0c6a2
Enable release notes translation

Releasenote translation publishing is being prepared. 'locale_dirs'
needs to be defined in conf.py to generate translated version of the
release notes.

Note that this repository might not get translated release notes - or
no translations at all - but we add the entry here nevertheless to
prepare for it.

Change-Id: I428d583eb3f3bbced321348e2a71035fbb282806