Updated openstack/openstack Project: openstack/neutron-lbaas  68ae7d9d2cf6b64b66bd9e76ed44919ec014fe53
Fix a gate issue with a mock and deprecated call

The change that relocates teh Agent DB model changed the path to
the "is_agent_down" method.  The deprecation means the happy path
still works, but our mock for that method failed with the old location.

The deprecation will be addressed in a later patch.

[1] https://review.openstack.org/#/c/330870/

Change-Id: I4ae7bfb5d7c5b33ae7501d133b2ce7f1fb7605c3