Updated openstack/openstack Project: openstack/requirements  2a91391d3ea6d7726c7afb2a7656ed4db7a257fe
Update keystoneauth to 2.14

Keystoneauth 2.14 provides updates to user agent handling that will
improve log debugging across OpenStack. This requires additions to the
libraries consuming keystoneauth. To do this we need to first bump the
minimum version so we know those features are present.

We meant to do that with I63cf0786a67c21fa1de6742ebed76781336f72e0 but
got the wrong version.

Change-Id: Ibd731f6c0b8b6a61a67b743a78c31fc4655fa0fc